refactor: 페이지네이션 관련 컴포넌트 수정
Related to: #35
