+++ src/components/ui/grid/GridView.tsx
@@ -1,29 +1,51 @@
-'use client';
-import { usePathname } from 'next/navigation';
-import GridItem from './GridItem';
+/**
+ * GridView 컴포넌트는 주어진 항목을 그리드 레이아웃으로 렌더링합니다.
+ *
+ * @template T 항목의 타입을 나타냅니다.
+ * @param {Object} props - 컴포넌트의 속성입니다.
+ * @param {T[]} props.items - 렌더링할 항목들의 배열입니다.
+ * @param {React.ComponentType<{ item: T }>} props.GridItem - 각 항목을 렌더링할 컴포넌트입니다.
+ * @param {'web4mobile2' | 'web3mobile1'} props.columnStyle - 모바일 및 웹에서의 그리드 열 수를 정의하는 스타일입니다.
+ *    - 'web4mobile2': 모바일에서 2열, 웹에서 4열 그리드
+ *    - 'web3mobile1': 모바일에서 1열, 웹에서 3열 그리드
+ * @param {'gapStyle1' | 'gapStyle2' | 'gapStyle3'} props.gapStyle - 그리드 항목 간의 간격 스타일입니다.
+ *    - 'gapStyle1': BEST PICK, 토마토들 추천활동, 토마토 Pick에 사용됩니다.
+ *    - 'gapStyle2': 대외활동, 공모전에 사용됩니다.
+ *    - 'gapStyle3': 토마토 Tip에 사용됩니다.
+ *
+ * @returns {JSX.Element} 그리드 레이아웃으로 렌더링된 항목들을 반환합니다.
+ */
 
-type GridViewProps = {
-  activities: Activity[];
-};
+export default function GridView<T>({
+  items,
+  GridItem,
+  columnStyle,
+  gapStyle,
+}: GridViewProps<T>) {
+  // 커스텀 gap 정의
 
-export default function GridView({ activities }: GridViewProps) {
-  const pathname = usePathname();
+  const columnStyles = {
+    web4mobile2: { mobile: 2, web: 4 },
+    web3mobile1: { mobile: 1, web: 3 },
+  };
 
-  // 공모전, 대외활동, 교육강연 페이지 여부 확인
-  const isSpecialRoute = ['/activity', '/contest'].includes(pathname);
+  const gapStyles = {
+    gapStyle1: 'gap-x-[14px] gap-y-[40px] md:gap-x-[20px] md:gap-y-[80px]', // BEST PICK, 토마토들 추천활동, 토마토 Pick
+    gapStyle2: 'gap-x-[14px] gap-y-[32px] md:gap-x-[20px] md:gap-y-[64px]', // 대외활동, 공모전
+    gapStyle3: 'gap-x-[0px] gap-y-[16px] md:gap-x-[30px] md:gap-y-[80px]', // 토마토 Tip
+  };
+
+  const gridClass = `
+    grid 
+    grid-cols-${columnStyles[columnStyle].mobile} 
+    md:grid-cols-${columnStyles[columnStyle].web} 
+    ${gapStyles[gapStyle]}
+  `.trim();
 
   return (
-    <div
-      className={
-        isSpecialRoute
-          ? // 공모전, 대외활동, 교육강연 페이지: PC 4x4, 모바일 2x8
-            'grid w-full grid-cols-2 gap-4 sm:grid-cols-2 md:grid-cols-4'
-          : // 그 외의 페이지: PC 4x2, 모바일 2x4
-            'grid w-full grid-cols-2 gap-4 sm:grid-cols-2 md:grid-cols-4'
-      }
-    >
-      {activities.map((activity, index) => (
-        <GridItem key={index} activity={activity} />
+    <div className={gridClass}>
+      {items.map((item, index) => (
+        <GridItem key={index} item={item} />
       ))}
     </div>
   );
